+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Bugfixes:
 * Handle variable spacing in MethylQA logs
 * Fixed bug in changing module search patterns, reported by @lweasel
 * Added timeout parameter to version check to prevent hang on systems with long defaults
+* Fixed `--exclude` so that it works with directories as well as files
 
 #### [v0.6](https://github.com/ewels/MultiQC/releases/tag/v0.6) - 2016-04-29
 Module updates: